requirements.txt
-i https://pypi.org/simple
certifi==2023.11.17; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
colorama==0.4.6; platform_system == 'Windows'
faiss-cpu==1.7.4
filelock==3.13.1; python_version >= '3.8'
fsspec==2023.12.2; python_version >= '3.8'
idna==3.6; python_version >= '3.5'
jinja2==3.1.3; python_version >= '3.7'
joblib==1.3.2; python_version >= '3.7'
markupsafe==2.1.3; python_version >= '3.7'
mpmath==1.3.0
networkx==3.2.1; python_version >= '3.9'
numpy==1.26.3; python_version >= '3.9'
pillow==10.2.0; python_version >= '3.8'
requests==2.31.0; python_version >= '3.7'
scikit-learn==1.3.2; python_version >= '3.8'
scipy==1.11.4; python_version >= '3.9'
sympy==1.12; python_version >= '3.8'
threadpoolctl==3.2.0; python_version >= '3.8'
torch==2.1.2; python_full_version >= '3.8.0'
torchvision==0.16.2; python_version >= '3.8'
tqdm==4.66.1; python_version >= '3.7'
typing-extensions==4.9.0; python_version >= '3.8'
urllib3==2.1.0; python_version >= '3.8'